chore: python kernel cleanup
