+++ src/inversify.config.ts
@@ -39,7 +39,7 @@ import { AltimateRequest } from "./altimate";
 import { ValidationProvider } from "./validation_provider";
 import { DeferToProdService } from "./services/deferToProdService";
 import { SharedStateService } from "./services/sharedStateService";
-import { NotebookKernelClient } from "./notebook_provider/notebookKernelClient";
+import { NotebookKernelClient } from "./notebook_provider/python/notebookKernelClient";
 import { NotebookDependencies } from "./notebook_provider/python/notebookDependencies";
 
 export const container = new Container();
