+++ src/notebook_provider/python/constants.ts
@@ -1,0 +1,21 @@
+export namespace Identifiers {
+  export const GeneratedThemeName = "ipython-theme"; // This needs to be all lower class and a valid class name.
+  export const MatplotLibDefaultParams = "_VSCode_defaultMatplotlib_Params";
+  export const MatplotLibFigureFormats = "_VSCode_matplotLib_FigureFormats";
+  export const DefaultCodeCellMarker = "# %%";
+  export const DefaultCommTarget = "jupyter.widget";
+  export const ALL_VARIABLES = "ALL_VARIABLES";
+  export const KERNEL_VARIABLES = "KERNEL_VARIABLES";
+  export const DEBUGGER_VARIABLES = "DEBUGGER_VARIABLES";
+  export const PYTHON_VARIABLES_REQUESTER = "PYTHON_VARIABLES_REQUESTER";
+  export const MULTIPLEXING_DEBUGSERVICE = "MULTIPLEXING_DEBUGSERVICE";
+  export const RUN_BY_LINE_DEBUGSERVICE = "RUN_BY_LINE_DEBUGSERVICE";
+  export const REMOTE_URI = "https://remote/";
+  export const REMOTE_URI_ID_PARAM = "id";
+  export const REMOTE_URI_HANDLE_PARAM = "uriHandle";
+  export const REMOTE_URI_EXTENSION_ID_PARAM = "extensionId";
+}
+
+export const WIDGET_MIMETYPE = "application/vnd.jupyter.widget-view+json";
+export const WIDGET_STATE_MIMETYPE =
+  "application/vnd.jupyter.widget-state+json";
